Addition of new FEA configuration statements to retain XORP unicast forwarding entries on startup or shutdown:
fea {
    unicast-forwarding4 {
        forwarding-entries {
            retain-on-startup: false
            retain-on-shutdown: false
        }
    }
    unicast-forwarding6 {
        forwarding-entries {
            retain-on-startup: false
            retain-on-shutdown: false
        }
    }
}

The default value for each statement is false.
Note that those statements prevent the FEA itself from deleting
the forwarding entries and does not prevent the RIB or any of the
unicast routing protocols from deleting the entries on shutdown.

This change partially fixes Bugzilla entry #665 and is needed
for implementing graceful restarting capabilities for the unicast
routing protocols.

Bugzilla URL:	http://www.xorp.org/bugzilla/show_bug.cgi?id=665


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@9313 26435e26-4b25-483a-942d-477236992e84

+++ trunk/xorp/fea/fticonfig_table_set_click.cc
@@ -12,7 +12,7 @@
 // notice is a summary of the XORP LICENSE file; the license in that file is
 // legally binding.
 
-#ident "$XORP: xorp/fea/fticonfig_table_set_click.cc,v 1.9 2005/03/25 02:53:05 pavlin Exp $"
+#ident "$XORP: xorp/fea/fticonfig_table_set_click.cc,v 1.10 2006/03/16 00:03:53 pavlin Exp $"
 
 #include "fea_module.h"
 
@@ -63,8 +63,11 @@ FtiConfigTableSetClick::start(string& error_msg)
     if (ClickSocket::start(error_msg) < 0)
 	return (XORP_ERROR);
 
-    delete_all_entries4();
-    delete_all_entries6();
+    // Cleanup any leftover entries from previously run XORP instance
+    if (! ftic().unicast_forwarding_entries_retain_on_startup4())
+	delete_all_entries4();
+    if (! ftic().unicast_forwarding_entries_retain_on_startup6())
+	delete_all_entries6();
 
     _is_running = true;
 
@@ -89,8 +92,11 @@ FtiConfigTableSetClick::stop(string& error_msg)
     if (! _is_running)
 	return (XORP_OK);
 
-    delete_all_entries4();
-    delete_all_entries6();
+    // Delete the XORP entries
+    if (! ftic().unicast_forwarding_entries_retain_on_shutdown4())
+	delete_all_entries4();
+    if (! ftic().unicast_forwarding_entries_retain_on_shutdown6())
+	delete_all_entries6();
 
     ret_value = ClickSocket::stop(error_msg);
