+++ trunk/xorp/fea/fticonfig_table_set_rtmv2.cc
@@ -12,7 +12,7 @@
 // notice is a summary of the XORP LICENSE file; the license in that file is
 // legally binding.
 
-#ident "$XORP: xorp/fea/fticonfig_table_set_rtsock.cc,v 1.13 2006/03/16 00:03:53 pavlin Exp $"
+#ident "$XORP: xorp/fea/fticonfig_table_set_rtmv2.cc,v 1.1 2006/06/29 11:03:55 bms Exp $"
 
 #include "fea_module.h"
 
@@ -60,8 +60,10 @@ FtiConfigTableSetRtmV2::start(string& error_msg)
 	return (XORP_OK);
 
     // Cleanup any leftover entries from previously run XORP instance
-    delete_all_entries4();
-    delete_all_entries6();
+    if (! ftic().unicast_forwarding_entries_retain_on_startup4())
+	delete_all_entries4();
+    if (! ftic().unicast_forwarding_entries_retain_on_startup6())
+	delete_all_entries6();
 
     //
     // XXX: This mechanism relies on the FtiConfigEntrySet mechanism
@@ -81,8 +83,11 @@ FtiConfigTableSetRtmV2::stop(string& error_msg)
     if (! _is_running)
 	return (XORP_OK);
 
-    delete_all_entries4();
-    delete_all_entries6();
+    // Delete the XORP entries
+    if (! ftic().unicast_forwarding_entries_retain_on_shutdown4())
+	delete_all_entries4();
+    if (! ftic().unicast_forwarding_entries_retain_on_shutdown6())
+	delete_all_entries6();
 
     //
     // XXX: This mechanism relies on the FtiConfigEntrySet mechanism
